requirements.txt
asn1crypto==1.5.1
boto3==1.38.0
botocore==1.38.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
confluent-kafka==2.10.0
cryptography==44.0.2
Faker==37.1.0
filelock==3.18.0
idna==3.10
jmespath==1.0.1
kafka-python==2.1.5
packaging==25.0
platformdirs==4.3.7
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
s3transfer==0.12.0
six==1.17.0
snowflake-connector-python==3.14.1
sortedcontainers==2.4.0
tomlkit==0.13.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
